Get rid of upload_all vs. upload_diff
Diffing against an empty tree will get a list of all blobs and
submodules of a repository. upload_diff on that diff will have the same
effect as upload_all. So upload_all can be removed.

You get an empty tree by invoking:

  git mktree -z < /dev/null

That can be hardcoded to:

  "4b825dc642cb6eb9a060e54bf8d69288fbee4904"

Thanks to panike from the #git IRC channel.